+++ src/utils/utils.ts
@@ -1,30 +1,22 @@
-export function calcPrincipal(interest: number, duration: number, endAmount: number): number {
-  // TODO: option to calculate interest which is e.g. compounded monthly:
-  // endAmount / Math.pow(1 + i / 12, duration * 12);
+export function calcPrincipal(interest: number, duration: number, endAmount: number, interval = 1): number {
   const i = interest / 100;
-  const principal = endAmount / Math.pow(1 + i, duration);
+  const principal = endAmount / Math.pow(1 + i / interval, duration * interval);
   return +principal.toFixed(2);
 }
 
-export function calcInterest(startAmount: number, duration: number, endAmount: number): number {
-  // TODO: option to calculate interest which is e.g. compounded monthly:
-  // 12 * (Math.pow(endAmount / startAmount, 1 / (12 * duration)) - 1);
-  const interest = Math.pow(endAmount / startAmount, 1 / duration) - 1;
+export function calcInterest(startAmount: number, duration: number, endAmount: number, interval = 1): number {
+  const interest = interval * (Math.pow(endAmount / startAmount, 1 / (interval * duration)) - 1);
   return +(interest * 100).toFixed(2);
 }
 
-export function calcDuration(startAmount: number, interest: number, endAmount: number): number {
-  // TODO: option to calculate interest which is e.g. compounded monthly:
-  // Math.log(endAmount / startAmount) / 12 * Math.log(1 + i / 12);
+export function calcDuration(startAmount: number, interest: number, endAmount: number, interval = 1): number {
   const i = interest / 100;
-  const duration = Math.log(endAmount / startAmount) / Math.log(1 + i);
+  const duration = Math.log(endAmount / startAmount) / (interval * Math.log(1 + i / interval));
   return +duration.toFixed(2);
 }
 
-export function calcFinalAmount(startAmount: number, interest: number, duration: number): number {
-  // TODO: option to calculate interest which is e.g. compounded monthly:
-  // startAmount * Math.pow(1 + i/12, duration*12);
+export function calcFinalAmount(startAmount: number, interest: number, duration: number, interval = 1): number {
   const i = interest / 100;
-  const endAmount = startAmount * Math.pow(1 + i, duration);
+  const endAmount = startAmount * Math.pow(1 + i / interval, interval * duration);
   return +endAmount.toFixed(2);
 }