+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0
+
+- it's now possible to set the symbol in next to the `principal` and `interest` fields
+- option to set interval of compounding
+- set correct licence (GPLv3) in `package.json`
+- updated dependencies
+
 ## 1.0.1
 
 - removed header from component